+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/handlers/ModulesSorter.java
@@ -2,7 +2,8 @@ package com.sap.cloud.lm.sl.mta.handlers;
 
 import java.util.List;
 
-import com.sap.cloud.lm.sl.mta.model.v1.Module;
+import com.sap.cloud.lm.sl.mta.model.v2.Module;
+
 
 public interface ModulesSorter {
